Rename `HttpHandler` to `HttpResponseHandler`
